# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IdeaExchange API",
    "version": "1.0.0",
    "description": "API documentation for the IdeaExchange platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Idea": {
        "type": "object",
        "required": [
          "title",
          "description",
          "author",
          "category",
          "department"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "author": {
            "type": "string",
            "description": "User ID of the author"
          },
          "coAuthors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of User IDs"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "in_review",
              "approved",
              "implemented",
              "rejected"
            ]
          },
          "category": {
            "type": "string",
            "description": "Category ID"
          },
          "department": {
            "type": "string"
          },
          "votes": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "currentVersion": {
            "type": "number"
          }
        }
      },
      "NewIdea": {
        "type": "object",
        "required": [
          "title",
          "description",
          "category",
          "department"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "description": "Category ID"
          },
          "department": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "role",
          "department"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "moderator",
              "admin",
              "department_head"
            ]
          },
          "department": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "points": {
            "type": "number"
          },
          "badges": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "role",
                  "department"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "moderator",
                      "admin",
                      "department_head"
                    ]
                  },
                  "department": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful registration"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login"
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout a user",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successful logout"
          }
        }
      }
    },
    "/ideas": {
      "get": {
        "summary": "Retrieve all ideas",
        "tags": [
          "Ideas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of ideas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Idea"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new idea",
        "tags": [
          "Ideas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewIdea"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created idea",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Idea"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}